pyproject.toml
[tool.poetry]
name = "ami"
version = "0.1.0"
description = ""
authors = ["bjablonski <akadrus@gmail.com>"]
[tool.poetry.dependencies]
python = "^3.6"
click = "^7.0"
pyyaml = "^5.3"
pyunpack = "^0.1.2"
patool = "^1.12"
tabulate = "^0.8.6"
[tool.poetry.dev-dependencies]
pytest = "^5.2"
[tool.poetry.scripts]
ami = 'ami:cli'
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"